Letting go of some test assertions.
I'm impressed with the level of testing ethic demonstrated, but copying
the significant body of the execute() method verbatim adds too much
maintenance burden for this little plugin that's really only used in one
instance.

If we are to add assertions on the content, it's probably suffice to
check that it rendered something and didn't fail.